requirements.txt
pinecone-client
openai
openai[embeddings]
pandas
tiktoken
python-dotenv
webvtt-py

